Add VCNL4200 UserDrivers to connect to framework
The VCNL4200 is a combination proximity and ambient light sensor, and
the new class allows users to register the driver with Sensor.PROXIMITY
and Sensor.LIGHT types with the framework. This change also fixes a
minor issue with bit alignment of the multi-pulse configuration.

The Vcnl4200.Configuration class has been introduced to help users
set all the attributes of the sensor they want, and provide a sane set
of defaults for attributes they aren't interested in. This class can be
passed as an input upon construction of the driver, or at any point
during the driver's lifetime. For UserDrivers, the Configuration class
must be provided in order to set the initial parameters of the sensor,
as users do not have access to the underlying I2cDevice it uses.

Bug: 65483431
Test: Validated against test application after registering sensors to
confirm that sensor events are generated and returned values are sane.
Change-Id: I8ada010ce140243d662f25f4df63ec7c380e693f